+++ tests/edit/merge/23.sh
@@ -1,0 +1,1 @@
+${PORTEDIT} merge -e 'IGNORE=foo' 23.in | diff -L 23.expected -L 23.actual -u 23.expected -